old_functions/requirements.txt
firebase-admin>=6.2.0
firebase-functions>=0.1.0
# Add other dependencies you need for video analysis:
# tensorflow
# opencv-python-headless>=4.8.0
# numpy>=1.24.0
# etc.
opencv-python-headless>=4.8.0
numpy>=1.24.0
mediapipe>=0.10.0
openai>=1.0.0
google-cloud-videointelligence>=2.11.1
google-cloud-vision>=3.4.0
asyncio>=3.4.3
python-dotenv>=1.0.0 